refactor(playground): 添加并发调度和错误处理工具，优化AlphaTex初始化流程
